% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_connection.R
\name{db_connection}
\alias{db_connection}
\title{Connecting to a database}
\usage{
db_connection(config_file = "./config/db_config.json", config_db)
}
\arguments{
\item{config_db}{The name of the database you would like to connect to in the config file.}
}
\value{
A database connection object.
}
\description{
Establishing a connection to a database.
#' @param config_file The path to a json config file. If none is given the default is "./config/db_config.json" in the current working directory. An example of the json file structure can be seen below.
}
\examples{
db_config = {
  "databases": {
    "db_name": {
      "host": "host.name.com",
      "port": 1234,
      "user": "user_name",
      "password": "pwd123",
      "db": "SID name"
    }
  }
}
}
