% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_string.R
\name{format_string}
\alias{format_string}
\title{Formatting a string with built in arguments}
\usage{
format_string(
  string,
  upper = TRUE,
  strip_punc = TRUE,
  strip_ws = TRUE,
  strip_comma = TRUE
)
}
\description{
Formatting a string with built in arguments
}
