% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison_functions.R
\name{\%!=na\%}
\alias{\%!=na\%}
\title{Compare values for inequality}
\usage{
v1 \%!=na\% v2
}
\value{
Boolean
}
\description{
Compare two vectors or factors but want comparisons with NA’s to be reported as TRUE or FALSE (instead of NA).
}
