% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_count.R
\name{row_count}
\alias{row_count}
\title{Counts the number of rows in a data object}
\usage{
row_count(.data)
}
\arguments{
\item{.data}{The data source.}
}
\value{
Number
}
\description{
This function can be used with a remote data source.
}
